scraper_bot/requirements.txt
playwright==1.50.0
aiohttp==3.11.13
beautifulsoup4==4.13.3
streamlit==1.43.1
faiss-cpu==1.10.0
ollama==0.4.7
fastapi==0.115.11
uvicorn==0.34.0